[GD32F1xx 누락 코드 업데이트] - drv namespace 삭제 누락 업데이트 - SDIO의 DMA 설정 오류 수정 - clock 관련 변수 선언을 unsigned int에서 int로 변경 - const unsigned int에서 const short로 변경

+++ yss/drv/clock/drv_clock_gd32f1.cpp
@@ -26,12 +26,12 @@
 
 #include <drv/Clock.h>
 
-unsigned int Clock::mHseFreq __attribute__((section(".non_init")));
-unsigned int Clock::mPllFreq __attribute__((section(".non_init")));
-unsigned int Clock::mLseFreq __attribute__((section(".non_init")));
+int Clock::mHseFreq __attribute__((section(".non_init")));
+int Clock::mPllFreq __attribute__((section(".non_init")));
+int Clock::mLseFreq __attribute__((section(".non_init")));
 
-static const unsigned int gPpreDiv[8] = {1, 1, 1, 1, 2, 4, 8, 16};
-static const unsigned int gHpreDiv[16] = {1, 1, 1, 1, 1, 1, 1, 1, 2, 4, 8, 16, 64, 128, 256, 512};
+static const short gPpreDiv[8] = {1, 1, 1, 1, 2, 4, 8, 16};
+static const short gHpreDiv[16] = {1, 1, 1, 1, 1, 1, 1, 1, 2, 4, 8, 16, 64, 128, 256, 512};
 
 bool Clock::enableHse(unsigned int hseHz, bool useOsc)
 {
